emerging/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.1.5
django-filter==2.2.0
django-cors-headers==2.4.0
djangorestframework==3.9.0
idna==2.8
mysqlclient==1.3.14
pytz==2019.3
requests==2.21.0